split models off
