+++ paddlenlp/prompt/__init__.py
@@ -12,9 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from .prompt_args import *
+from .prompt_model import *
+from .prompt_tokenizer import *
+from .prompt_trainer import *
+from .prompt_utils import *
 from .template import *
 from .verbalizer import *
-from .prompt_args import *
-from .prompt_trainer import *
-from .prompt_tokenizer import *
-from .prompt_utils import *
